change the rating sturcture

+++ database/migrations/2024_03_29_220051_add_freshness_to_ratings.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('ratings', function (Blueprint $table) {
-            $table->float('total')->change();
+            $table->tinyInteger('total')->change();
         });
     }
